requirements.txt
motor==3.5.1
pymongo==4.8.0
python-dotenv==1.0.1
aio-pika==9.4.3
pydantic==2.8.2
fastapi==0.111.0